+++ post-types/templates/single-docs.php
@@ -20,9 +20,9 @@ get_header(); ?>
     <article <?php post_class(); ?> id="post-<?php the_ID(); ?>">
     	<div class="post_content mybooking-docs">
     		<div class="container" tabindex="-1">
-    			<div class="mb-row">
+    			<div class="mb-row row">
 
-    				<div class="mb-col-md-10 mb-col-center">
+    				<div class="mb-col-md-10 col-md-12 mb-col-center">
 
 							<?php echo mybooking_breadcrumbs(); ?>
 
@@ -65,8 +65,8 @@ get_header(); ?>
 
 							<!-- Categories navigation -->
 
-							<div class="mybooking-docs_footer-nav">
-								<div class="mb-col-md-6">
+							<div class="mybooking-docs_footer-nav mb-row row">
+								<div class="mb-col-md-6 col-md-6">
 
 									<!-- Help categories -->
 									<h3 class="mybooking-docs_categories-title">
@@ -96,7 +96,7 @@ get_header(); ?>
 									</ul>
 								</div>
 
-								<div class="mb-col-md-6">
+								<div class="mb-col-md-6 col-md-6">
 
 									<!-- Docs categories -->
 									<h3 class="mybooking-docs_categories-title">