+++ CHANGELOG.md
@@ -1,12 +1,32 @@
-# Mybooking Docs plugin Changelog
+# Mybooking Docs Plugin Changelog
 
-## 1.0.3
+## 1.0.7
+
+### Fixed
+- Mybooking Theme columns compatibility
+
+## 1.0.6
 
 ### Refactor
 - Moved CTP code to separated folder
 
 ### Added
 - Portfolio CPT and custom template
+
+## 1.0.5
+
+### Fixed
+- Archive page
+
+## 1.0.4
+
+### Fixed
+- Grid system : Use bootstrap 4
+
+## 1.0.3
+
+### Fixed
+- Removed terms
 
 ## 1.0.2